requirements.txt
aiogram>=3.11.0
aiogram-dialog>=2.2.0
python-dotenv>=1.0.0